refactor(PoF): rename data model modules
* Rename prgoram DB model files to all lowercase.
